requirements.txt
strands-agents
boto3
requests
python-dotenv
tabulate
colorama
tqdm
